refactor dropdown
- using initial PR dropdownPlus.
- remove dropdownPlus;
